Simplify CMake build for portability.
Include http_parser in third_party since it does not
have a CMake buildfile.

Fetch abseil at build time since it's not in most third-party
repositories and we always link it statically as well, and because
the CMake build is the most usable.

Use the same mechanism to build gtest using CMake.